Issue #OS-36  PR comments addressed minnor refactoring
