+++ java/registry/src/main/resources/application.yml.sample
@@ -37,7 +37,7 @@ search:
   #   If NativeSearchService, then every search API uses the same database as the writes. May 
   # not offer high speed reads. This is the default search service, if this config is not provided.
   #   If ElasticSearchService, then Elastic search is used.
-  advisor: io.opensaber.registry.service.impl.NativeSearchService  
+  provider: io.opensaber.registry.service.NativeSearchService  
         
 database:
   # This property is internal and not to be confused with the schema definition.
@@ -59,7 +59,7 @@ database:
   # This property is instruction to use the shard advisor.
   # Values could be DefaultShardAdvisor, SerialNumberShardAdvisor
   # If this property not provided, advisor is set to DefaultShardAdvisor
-  shardAdvisor: SerialNumberShardAdvisor
+  shardAdvisor: io.opensaber.registry.sink.shard.SerialNumberShardAdvisor
 
   connectionInfo:
     -
